Rename bot.js to main.js
